+++ lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:todoapp/homescreen.dart';
+import 'package:todoapp/pages/homescreen.dart';
 
 void main() {
   runApp(const MyApp(),);
@@ -11,7 +11,7 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return const MaterialApp(
-      title: 'SeedScan',
+      title: 'TodoApp',
       debugShowCheckedModeBanner: false,
       home:  Homescreen(),
     );
