--- lib/task_details.dart
@@ -1,38 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:todoapp/const/const.dart';
-
-class TaskDetailsPage extends StatelessWidget {
-  final Map<String, dynamic> task;
-
-  const TaskDetailsPage({super.key, required this.task});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Task Details'),
-        backgroundColor: headerColor,
-      ),
-      body: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Text(
-              task['title'],
-              style: const TextStyle(
-                fontWeight: FontWeight.bold,
-                fontSize: 24,
-              ),
-            ),
-            const SizedBox(height: 10),
-            Text(
-              task['description'],
-              style: const TextStyle(fontSize: 18),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
